fix: improve network list item loading, fixes #1717
